% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malschains.R
\name{print.malschains}
\alias{print.malschains}
\title{Generic print function for malschains results}
\usage{
\method{print}{malschains}(x, ...)
}
\arguments{
\item{x}{the \code{\link{malschains}} result}

\item{...}{additional function parameters (currently not used)}
}
\description{
Print out some characteristics of a \code{\link{malschains}} result.
The result shows besides the best solution and its fitness the total number of evaluations spent for both EA and LS, 
the ratio of the spent evaluations (also called effort), the ratio of total improvement of the fitness,
the percentage of times that application of the EA/LS yielded improvement, and some timing results in milliseconds.
}
